Adding additional tooling
I was originally considering shimming Martin Thomson's tools in, however that
was a bit problematic so the plan is just use the various pieces (xml2rfc et al)
directly. For now this can generate the necessary XML, future commits should be
able to generate the HTML output and do idnits etc.

One thing in the design of this that I hold strong views on is decoupling it
from version control as much as possible, and having the least amount of code
committed with the drafts themselves - this will give me a bit more flexibility
as to how the tooling might be integrated with other things in the future like
using some sort of CI automation and the likes.